requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic>=2.8.2
pydantic-settings==2.1.0
email-validator==2.1.0
chromadb==0.4.22
mistralai>=1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
redis==5.0.1
numpy<2.0.0
